requirement.txt
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1
idna==2.10
jwcrypto==0.8
pycparser==2.20
pycryptodome==3.9.8
requests==2.24.0
six==1.15.0
urllib3==1.25.10
